backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
elasticsearch==8.17.0
openai==1.68.0
httpx==0.27.2
python-dotenv==1.0.0
requests==2.32.0
python-multipart==0.0.9